Add files via upload
There are code examples (the output was too large to upload) from three seperate projects here, all in scientific modelling. The first was for my diploma in theoretical physics, it is entirely written in c. The second is the record of the work i did for my Bsc Physics final year project, where i used latex to typeset the document, and some pythan and bash terminal coding to do the the actual astrophysical research. Finally, my masters by research project is uploaded, which consists of a backend of c, along with a wrapper of python. please email me at mo.b.awan.uk@gmail.com if you have any further questions. your, MBA.